Finished X notification work and git-ignored files.
