+++ .gitignore
@@ -1,8 +1,6 @@
 .gradle
 /local.properties
-/.idea/workspace.xml
-/.idea/libraries
+/.idea/**
 .DS_Store
-/build
-build/*
+**/build/**
 *.iml
