--- .idea/gradle.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="GradleSettings">
-    <option name="linkedExternalProjectsSettings">
-      <GradleProjectSettings>
-        <option name="distributionType" value="LOCAL" />
-        <option name="externalProjectPath" value="$PROJECT_DIR$" />
-        <option name="gradleHome" value="C:\Program Files\Android\Android Studio\gradle\gradle-2.14.1" />
-        <option name="modules">
-          <set>
-            <option value="$PROJECT_DIR$" />
-            <option value="$PROJECT_DIR$/Android-Terminal-Emulator" />
-            <option value="$PROJECT_DIR$/Android-Terminal-Emulator/emulatorview" />
-            <option value="$PROJECT_DIR$/Android-Terminal-Emulator/libtermexec" />
-            <option value="$PROJECT_DIR$/Android-Terminal-Emulator/term" />
-            <option value="$PROJECT_DIR$/GNURootDebianSource" />
-            <option value="$PROJECT_DIR$/bVNC" />
-            <option value="$PROJECT_DIR$/downloader_library" />
-            <option value="$PROJECT_DIR$/play_licensing" />
-            <option value="$PROJECT_DIR$/pubkeyGenerator" />
-            <option value="$PROJECT_DIR$/zoomerWithKeys" />
-          </set>
-        </option>
-        <option name="resolveModulePerSourceSet" value="false" />
-      </GradleProjectSettings>
-    </option>
-  </component>
-</project>
